# packages/api-client/openapi.yaml
openapi: 3.1.0
info:
  title: LitMood API
  description: 책·영화·음악을 무드와 함께 기록하고 나누는 서비스의 API
  version: v1
servers:
  - url: http://localhost:8080
    description: local
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/users/@{handle}:
    get:
      tags:
        - Users
      summary: 공개 프로필 조회
      operationId: profile
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PublicProfile'
      security: []
  /api/v1/users/@{handle}/records:
    get:
      tags:
        - Users
      summary: 공개 기록 목록
      description: PUBLIC 기록만 노출한다
      operationId: publicRecords
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
        - name: types
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
              enum:
                - BOOK
                - MOVIE
                - MUSIC
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RecordPage'
      security: []
  /api/v1/users/@{handle}/collections:
    get:
      tags:
        - Users
      summary: 공개 컬렉션 목록
      description: 본인이 조회하면 비공개 컬렉션도 포함된다
      operationId: collections
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollectionSummary'
      security: []
...
components:
  schemas:
...
    ContentRef:
      type: object
      description: 기록에 붙는 콘텐츠 요약
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum:
            - BOOK
            - MOVIE
            - MUSIC
        provider:
          type: string
          enum:
            - NAVER_BOOK
            - TMDB
            - SPOTIFY
        externalId:
          type: string
        title:
          type: string
        creators:
          type: array
          items:
            type: string
        releasedOn:
          type: string
          format: date
        coverUrl:
          type: string
    MoodTag:
      type: object
      properties:
        name:
          type: string
        displayName:
          type: string
        color:
          type: string
        curated:
          type: boolean
    RecordResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
          enum:
            - WANT
            - DOING
            - DONE
            - DROPPED
        rating:
          type: number
        moods:
          type: array
          items:
            $ref: '#/components/schemas/MoodTag'
        review:
          type: string
        isSpoiler:
          type: boolean
        visibility:
          type: string
          enum:
            - PUBLIC
            - FOLLOWERS
            - PRIVATE
        contextNote:
          type: string
        startedAt:
          type: string
          format: date
        finishedAt:
          type: string
          format: date
        repeatCount:
          type: integer
          format: int32
        likeCount:
          type: integer
          format: int32
        likedByMe:
          type: boolean
        authorHandle:
          type: string
        authorNickname:
          type: string
        content:
          $ref: '#/components/schemas/ContentRef'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PublicProfile:
      type: object
      properties:
        handle:
          type: string
        nickname:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        followers:
          type: integer
          format: int64
        following:
          type: integer
          format: int64
        followedByMe:
          type: boolean
    RecordPage:
      type: object
      description: 커서 기반 페이지. nextCursor 가 null 이면 마지막 페이지다.
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RecordResponse'
        nextCursor:
          type: string
        totalCount:
          type: integer
          format: int64
    CollectionSummary:
      type: object
      description: 목록용 — 아이템 없이 개수만
      properties:
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        coverUrl:
          type: string
        visibility:
          type: string
          enum:
            - PUBLIC
            - FOLLOWERS
            - PRIVATE
        itemCount:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT